requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
schedule==1.2.0
python-multipart==0.0.6
jinja2==3.1.2
